app/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
psycopg2-binary==2.9.7
pandas>=2.2.0
numpy>=1.24.0
scikit-learn>=1.3.0
mlxtend>=0.22.0
plotly>=5.17.0
python-dotenv==1.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

